refactor: break apart functions
