+++ tests/devpool-issue-handler.test.ts
@@ -1,7 +1,13 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { drop } from "@mswjs/data";
 import { setupServer } from "msw/node";
-import { calculateStatistics, checkIfForked, getPartnerUrls, getRepoUrls, GitHubIssue, newDirectoryIssue, syncIssueMetaData } from "../helpers/directory";
+import { calculateStatistics } from "../helpers/directory/calculate-statistics";
+import { checkIfForked } from "../helpers/directory/check-if-forked";
+import { GitHubIssue } from "../helpers/directory/directory";
+import { getPartnerUrls } from "../helpers/directory/get-partner-urls";
+import { getRepoUrls } from "../helpers/directory/get-repo-urls";
+import { newDirectoryIssue } from "../helpers/directory/new-directory-issue";
+import { syncIssueMetaData } from "../helpers/directory/sync-issue-meta-data";
 import { db } from "../mocks/db";
 import { handlers } from "../mocks/handlers";
 import issueDevpoolTemplate from "../mocks/issue-devpool-template.json";
